pagination - css override, moved all JS files from 'components' to 'js' folder
