+++ src/sass/index.scss
@@ -4,6 +4,7 @@
 
 @import './components/buttons';
 @import './components/movie-modal';
+@import './components/pagination';
 @import './components/team-modal';
 
 @import './layout/home-header';
